Update ag_ancient_knowledge_l_english.yml
Line 76 may need some closer attention. Look over it and make sure it's correct. The translation is fine, but i am concerned about the latter portion of the structure and how it will show up in-game.

Check line 88. I chose to translate this as "We can do research on the ancient ships we have.", but it could also be read as "We can do research on existing ancient ships" or perhaps "We can research the ancient ships still in existence". That last one is stretching it a bit, but i am putting it out there as a potential meaning. The original text is 我们可以对我们现有的古代舰船进行研究。

Line 117 may be Reverse Engineer or Reverse Engineering. I am unsure of the exact context it is used in. I believe it should be the latter, and have set it as such.

Line 127~140 do not have a space between ! and § (between Research§! and $Ancient) in the original files. These were in the existing english files, so i left them in and adjusted my additions correspondingly, but check if that is meant to be there. If it doesn't matter, ignore this.